django_blog/requirements.txt
django
django-taggit



# emmycon
# !5PNLUX#eAAb-ug


# hello12
# rcYk9!WnqVkTm@B